increasing numerical stability of gaussian approximations
Default optim() choice was BFGS, which is too aggressive of an algorithm near parameter boundaries, at which the eigendecompositions for the model become numerically unstable.  Switching to Nelder-Mead as the default choice helps stabilize the algorithm by focusing more on local searches.  The fix is not a guarantee, but seems to help in practice.